feat: start implementing logic of  getting info about a product by its id

+++ services/backend/product-service/src/main/java/com/veteroch4k/product/repositories/DrawingRepository.java
@@ -3,6 +3,6 @@ package com.veteroch4k.product.repositories;
 import com.veteroch4k.product.models.Drawing;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DrawingRepository extends JpaRepository<Drawing, Long> {
+public interface DrawingRepository extends JpaRepository<Drawing, Integer> {
 
 }
